+++ engine/src/execution_graph/logic_controllers/LogicPrimitives.cpp
@@ -121,7 +121,7 @@ std::vector<std::string> BlazingTableView::names() const{
 	return this->columnNames;
 }
 
-std::vector<cudf::data_type> BlazingTableView::get_schema() {
+std::vector<cudf::data_type> BlazingTableView::get_schema() const {
 	std::vector<cudf::data_type> data_types(this->num_columns());
 	auto view = this->view();
 	std::transform(view.begin(), view.end(), data_types.begin(), [](auto & col){ return col.type(); });